Move access token type enum to AccessTokenImpl
It's actually an implementation detail and should not be in the
interface.
